Refactor code and show components in playground
