+++ coveralls.json
@@ -1,9 +1,10 @@
 {
   "coverage_options": {
-    "minimum_coverage": 85.3
+    "minimum_coverage": 85.7
   },
   "skip_files": [
     "lib/omedis/release.ex",
-    "lib/omedis_web/components/core_components.ex"
+    "lib/omedis_web/components/core_components.ex",
+    "lib/omedis_web/components/calendar_components.ex"
   ]
 }
